--- ar.json
@@ -1,1 +1,0 @@
-{"username":"--"}
