+++ src/cli/cli.ts
@@ -1,10 +1,8 @@
-import { getLanguages, LanguageCodes, Sources } from '..';
+import { getLanguages,  Sources } from '..';
 import { fileTranslator, getFileFromPath } from '../core/json_file';
 import {
   error,
   info,
-  infoChoices,
-  language_choices,
   messages,
   success,
   warn,
@@ -12,8 +10,8 @@ import {
 import loading from 'loading-cli';
 import { capitalize, current_version, getCodeFromLanguage, translationStatistic } from '../utils/micro';
 import { readProxyFile } from '../core/proxy_file';
-var inquirer = require('inquirer');
 import { Command } from "commander";
+import { promptFrom, promptTo, promptTranslator } from '../utils/prompt';
 var figlet = require("figlet");
 
 const program = new Command();
@@ -47,16 +45,15 @@ export async function initializeCli() {
   */
   if (program.args[1] !== undefined && !program.args[1].includes('.txt')) {
     error(messages.cli.proxy_File_notValid_or_not_empty_options)
-    process.exit()
+    process.exit(1)
   }
 
 
   if (!process.argv.slice(2).length) {
     help();
     return;
-  } else
-
-    translate();
+  }
+  translate();
 
 }
 
@@ -92,132 +89,63 @@ async function translate() {
     error(messages.file.no_file_in_path);
     return;
   }
-  
-  
+
+
   let translator = myOpts.translator
   if (translator && translator !== '') {
     translator = capitalize(myOpts.translator)
     switch (translator) {
       case 'Google':
         global.source = Sources.GoogleTranslate
-
-        infoChoices(messages.cli.from_source, Sources.GoogleTranslate)
         break;
       case 'Argos':
         global.source = Sources.ArgosTranslate
-        infoChoices(messages.cli.from_source, Sources.ArgosTranslate)
 
         break;
       case 'Libre':
         global.source = Sources.LibreTranslate
-        infoChoices(messages.cli.from_source, Sources.LibreTranslate)
 
         break;
       case 'Bing':
         global.source = Sources.BingTranslate
-        infoChoices(messages.cli.from_source, Sources.BingTranslate)
         break;
       default:
         error(translator + ' ' + messages.cli.translator_not_available);
-        const source_choices = Object.entries(Sources).map(([key, _]) => {
-          return {
-            name: language_choices[key],
-            value: key,
-            short: key,
-          };
-        });
-        await inquirer
-          .prompt([
-            {
-              type: 'list',
-              name: 'source',
-              message: messages.cli.from_source,
-              pageSize: 20,
-              choices: [...source_choices, new inquirer.Separator()],
-            },
-          ])
-          .then((answers: any) => {
-            global.source = answers.source;
-          });
+        promptTranslator()
         break;
     }
   } else {
-    const source_choices = Object.entries(Sources).map(([key, _]) => {
-      return {
-        name: language_choices[key],
-        value: key,
-        short: key,
-      };
-    });
-    await inquirer
-      .prompt([
-        {
-          type: 'list',
-          name: 'source',
-          message: messages.cli.from_source,
-          pageSize: 20,
-          choices: [...source_choices, new inquirer.Separator()],
-        },
-      ])
-      .then((answers: any) => {
-        global.source = answers.source;
-      });
+   await promptTranslator()
   }
 
 
-  async function promptFrom(from_choices: any[], messages: any) {
-    const answers = await inquirer.prompt([
-      {
-        type: 'list',
-        name: 'from',
-        message: messages.cli.from_message,
-        pageSize: 20,
-        choices: [...from_choices, new inquirer.Separator()],
-      },
-    ]);
 
-    return answers;
-  }
-
-  async function promptTo(to_choices: any[], messages: any, default_languages?: any) {
-    const answers = await inquirer.prompt([
-      {
-        type: 'checkbox',
-        name: 'to',
-        pageSize: 20,
-        message: messages.cli.to_message,
-        choices: to_choices,
-        default: default_languages ? default_languages : []
-      },
-    ]);
-    return answers;
-  }
 
   const opt_from: any = myOpts.from ? capitalize(myOpts.from) : undefined;
   const opt_to: any = myOpts.to ? (myOpts.to as string).split(',').map(lang => capitalize(lang)) : undefined;
   let from!: string;
   let to!: string[];
-  const { from_choices, to_choices } = getLanguageChoices();
+  
 
   if (!opt_from && !opt_to) {
-    ({ from } = await promptFrom(from_choices, messages));
-    ({ to } = await promptTo(to_choices, messages));
+    ({ from } = await promptFrom());
+    ({ to } = await promptTo());
   } else {
     if (opt_from) {
       if ((getLanguages() as any)[opt_from]) {
         from = opt_from;
       } else {
         error(opt_from + ' ' + messages.cli.from_not_available);
-        ({ from } = await promptFrom(from_choices, messages));
+        ({ from } = await promptFrom());
       }
     } else {
-      ({ from } = await promptFrom(from_choices, messages));
+      ({ from } = await promptFrom());
     }
 
     if (opt_to) {
       to = opt_to;
     } else {
-      ({ to } = await promptTo(to_choices, messages));
+      ({ to } = await promptTo());
     }
   }
 
@@ -228,14 +156,14 @@ async function translate() {
     for (let lang of not_available_languages) {
       error(lang + ' ' + messages.cli.from_not_available);
     }
-    ({ to } = await promptTo(to_choices, messages, available_languages));
+    ({ to } = await promptTo(available_languages));
   }
 
 
 
   if (to.length === 0 || to === undefined) {
     warn(messages.cli.no_selected_language);
-    ({ to } = await promptTo(to_choices, messages));
+    ({ to } = await promptTo());
   }
 
   const to_languages = to.map(language => (getLanguages() as any)[language]);
@@ -271,18 +199,4 @@ async function translate() {
   info(messages.cli.creation_done);
 }
 
-function getLanguageChoices(): {
-  from_choices: LanguageCodes;
-  to_choices: LanguageCodes;
-} {
-  let from_choices = getFromChoices();
-  let to_choices = from_choices.filter(language => language !== `Automatic`);
 
-  return { from_choices, to_choices };
-}
-
-function getFromChoices(): LanguageCodes {
-  let languages = Object.entries(getLanguages() as any).map(([key, _]) => key);
-
-  return languages;
-}
